rename errorHelper dir
