+++ src/errorHelper/resultType.ts
